refactor: 단어장 entity 변경으로 인한folder entity 삭제
