+++ module-api/src/main/java/com/numo/api/domain/quiz/dto/stat/QuizStatResponseDto.java
@@ -9,8 +9,6 @@ import java.time.LocalDateTime;
 @Builder
 public record QuizStatResponseDto(
         Long quizStatId,
-        @Deprecated
-        String folderName,
         String wordBookName,
         int totalCount,
         int correctCount,
@@ -19,7 +17,7 @@ public record QuizStatResponseDto(
         String updateTime
 ) {
     public QuizStatResponseDto(Long quizStatId, String wordBookName, int totalCount, int correctCount, int wrongCount, LocalDateTime createTime, LocalDateTime updateTime) {
-        this(quizStatId, wordBookName, wordBookName, totalCount, correctCount, wrongCount, Timestamped.getFormatTime(createTime, "yyyy-MM-dd hh:mm:ss"), Timestamped.getFormatTime(updateTime, "yyyy-MM-dd hh:mm:ss"));
+        this(quizStatId, wordBookName, totalCount, correctCount, wrongCount, Timestamped.getFormatTime(createTime, "yyyy-MM-dd hh:mm:ss"), Timestamped.getFormatTime(updateTime, "yyyy-MM-dd hh:mm:ss"));
     }
 
     public static QuizStatResponseDto of(QuizStat quizStat) {
