--- module-api/src/main/java/com/numo/api/domain/wordbook/folder/dto/read/FolderInWordCountDto.java
@@ -1,8 +1,0 @@
-package com.numo.api.domain.wordbook.folder.dto.read;
-
-@Deprecated
-public record FolderInWordCountDto(
-        Long folderId,
-        Long count
-) {
-}
