--- module-api/src/main/java/com/numo/api/domain/wordbook/folder/dto/read/FolderListReadResponseDto.java
@@ -1,10 +1,0 @@
-package com.numo.api.domain.wordbook.folder.dto.read;
-
-import com.numo.api.domain.wordbook.folder.dto.FolderResponseDto;
-
-@Deprecated
-public record FolderListReadResponseDto(
-        FolderResponseDto folders,
-        Long count
-) {
-}
