+++ module-api/src/main/java/com/numo/api/domain/wordbook/word/aop/WordAspect.java
@@ -3,6 +3,7 @@ package com.numo.api.domain.wordbook.word.aop;
 import com.numo.api.domain.dictionary.dto.DictionaryDto;
 import com.numo.api.domain.dictionary.service.DictionaryCacheService;
 import com.numo.api.domain.dictionary.service.DictionaryService;
+import com.numo.api.domain.wordbook.sound.service.SoundService;
 import com.numo.api.domain.wordbook.word.dto.WordResponseDto;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.annotation.AfterReturning;
@@ -16,11 +17,13 @@ public class WordAspect {
 
     private final DictionaryService dictionaryService;
     private final DictionaryCacheService dictionaryCacheService;
+    private final SoundService soundService;
 
     public WordAspect(DictionaryService dictionaryService,
-                      DictionaryCacheService dictionaryCacheService) {
+                      DictionaryCacheService dictionaryCacheService, SoundService soundService) {
         this.dictionaryService = dictionaryService;
         this.dictionaryCacheService = dictionaryCacheService;
+        this.soundService = soundService;
     }
 
     /**
@@ -36,7 +39,8 @@ public class WordAspect {
 
         DictionaryDto dict = dictionaryService.save(dictionaryDto);
         // 캐시 저장소에도 이미 사전 데이터베이스에 있거나, 제대로된 단어 데이터가 아니면 저장하지 않음
-        if (!dict.mean().isEmpty()) {
+        System.out.println("dict: " + dict.isRealWord());
+        if (dict.isRealWord()) {
             dictionaryCacheService.save("dict", res.word());
         }
     }
