+++ module-api/src/main/java/com/numo/api/domain/wordbook/word/dto/WordRequestDto.java
@@ -7,13 +7,16 @@ import com.numo.domain.wordbook.detail.WordDetail;
 import com.numo.domain.wordbook.sound.Sound;
 import com.numo.domain.wordbook.sound.type.GttsCode;
 import com.numo.domain.wordbook.word.Word;
+import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 
 import java.util.List;
 
 @Builder
 public record WordRequestDto(
-        Long folderId,
+        @NotNull
+        Long wordBookId,
+        @NotNull
         String word,
         String mean,
         String read,
@@ -21,46 +24,24 @@ public record WordRequestDto(
         String memorization,
         List<WordDetailRequestDto> details
 ) {
-
-    public Word toEntity(Long userId, String gttsType) {
-        User user = User.builder().userId(userId).build();
-        WordBook folder = WordBook.builder().id(folderId).build();
-        List<WordDetail> wordDetails = details.stream().map(WordDetailRequestDto::toEntity).toList();
-        return Word.builder()
-                .user(user)
-                .wordbook(folder)
-                .wordDetails(wordDetails)
-                .word(word)
-                .mean(mean)
-                .read(read)
-                .memo(memo)
-                .memorization(memorization)
-                .lang(GttsCode.valueOf(gttsType))
-                .build();
+    public WordRequestDto {
+        word = word.replaceAll("\\s", "");
     }
 
-    public Word toEntity(Long userId, String gttsType, Long soundId) {
-        User user = User.builder().userId(userId).build();
-        WordBook folder = null;
-        if (folderId != null) {
-            folder = WordBook.builder().id(folderId).build();
-        }
-        Sound sound = null;
-        if (soundId != null) {
-            sound = Sound.builder().soundId(soundId).build();
-        }
+    public Word toEntity(User user, Sound sound, WordBook wordBook, GttsCode gttsType) {
         List<WordDetail> wordDetails = details.stream().map(WordDetailRequestDto::toEntity).toList();
         return Word.builder()
                 .user(user)
-                .wordbook(folder)
-                .wordDetails(wordDetails)
-                .word(word)
-                .mean(mean)
-                .read(read)
-                .memo(memo)
-                .memorization(memorization)
-                .lang(GttsCode.valueOf(gttsType))
                 .sound(sound)
+                .wordbook(wordBook)
+                .wordDetails(wordDetails)
+                .word(word)
+                .mean(mean)
+                .read(read)
+                .memo(memo)
+                .memorization(memorization)
+                .lang(gttsType)
                 .build();
     }
+
 }
