+++ module-api/src/main/java/com/numo/api/domain/wordbook/word/dto/read/ReadWordResponseDto.java
@@ -10,8 +10,6 @@ import java.util.List;
 @Builder
 public record ReadWordResponseDto(
         Long wordId,
-        @Deprecated
-        Long folderId,
         Long wordBookId,
         String word,
         String mean,
@@ -26,7 +24,6 @@ public record ReadWordResponseDto(
         public static ReadWordResponseDto of(WordDto word, List<ReadWordDetailListResponseDto> details) {
                 return ReadWordResponseDto.builder()
                         .wordId(word.wordId())
-                        .folderId(word.folderId())
                         .wordBookId(word.wordBookId())
                         .soundPath(word.soundPath())
                         .word(word.word())
