+++ module-api/src/main/java/com/numo/api/domain/wordbook/word/dto/search/SearchWordRequestDto.java
@@ -3,8 +3,6 @@ package com.numo.api.domain.wordbook.word.dto.search;
 import com.numo.domain.wordbook.type.SortType;
 
 public record SearchWordRequestDto(
-        @Deprecated
-        Long folderId,
         Long wordBookId,
         Long lastWordId,
         String memorization,
@@ -12,9 +10,4 @@ public record SearchWordRequestDto(
         SortType sort,
         String seed
 ) {
-    public SearchWordRequestDto {
-        if (folderId != null) {
-            wordBookId = folderId;
-        }
-    }
 }
