+++ module-api/src/main/java/com/numo/api/domain/wordbook/word/repository/query/WordCustomRepository.java
@@ -1,15 +1,13 @@
 package com.numo.api.domain.wordbook.word.repository.query;
 
-import com.numo.api.domain.wordbook.folder.dto.read.FolderInWordCountDto;
 import com.numo.api.domain.dailySentence.dto.DailyWordListDto;
-import com.numo.api.domain.wordbook.word.dto.read.ReadWordRequestDto;
-import com.numo.api.domain.wordbook.word.dto.WordDto;
 import com.numo.api.domain.wordbook.detail.dto.WordDetailResponseDto;
+import com.numo.api.domain.wordbook.word.dto.WordDto;
+import com.numo.api.domain.wordbook.word.dto.read.ReadWordRequestDto;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
 import java.util.List;
-import java.util.Map;
 
 public interface WordCustomRepository {
     /**
