--- module-domain/src/main/java/com/numo/domain/wordbook/folder/dto/FolderUpdateDto.java
@@ -1,13 +1,0 @@
-package com.numo.domain.wordbook.folder.dto;
-
-import lombok.Builder;
-
-@Builder
-@Deprecated
-public record FolderUpdateDto(
-        String folderName,
-        String memo,
-        String color,
-        String background
-) {
-}
