+++ module-domain/src/main/java/com/numo/domain/wordbook/type/UpdateType.java
@@ -1,5 +1,5 @@
 package com.numo.domain.wordbook.type;
 
 public enum UpdateType {
-    all, memo, memorization, @Deprecated folder
+    all, memo, memorization
 }
